// src/protocols/mrp/generated/DeviceInfoMessage.ts
export enum PreferredEncoding_Enum {
  Default = 0,
  JSON = 1,
  UNRECOGNIZED = -1,
}
// ...
export interface DeviceInfoMessage {
  /** Example: B8D8678C-9DA9-4D29-9338-5D6B827B8063 */
  uniqueIdentifier?: string | undefined;
  /** Example: Jean's iPhone */
  name: string;
  /** Example: iPhone */
  localizedModelName?: string | undefined;
  /** Example: 13F69 */
  systemBuildVersion?: string | undefined;
  /** Example: com.example.myremote */
  applicationBundleIdentifier?: string | undefined;
  /** Example: 107 */
  applicationBundleVersion?: string | undefined;
  /** Example: 1 */
  protocolVersion?: number | undefined;
  lastSupportedMessageType?: number | undefined;
  supportsSystemPairing?: boolean | undefined;
  allowsPairing?: boolean | undefined;
  connected?: boolean | undefined;
  systemMediaApplication?: string | undefined;
  supportsACL?: boolean | undefined;
  supportsSharedQueue?: boolean | undefined;
  supportsExtendedMotion?: boolean | undefined;
  bluetoothAddress?: Buffer | undefined;
  sharedQueueVersion?: number | undefined;
  deviceUID?: string | undefined;
  managedConfigDeviceID?: string | undefined;
  deviceClass?: DeviceClass_Enum | undefined;
  logicalDeviceCount?: number | undefined;
  tightlySyncedGroup?: boolean | undefined;
  isProxyGroupPlayer?: boolean | undefined;
  tightSyncUID?: string | undefined;
  groupUID?: string | undefined;
  groupName?: string | undefined;
  groupedDevices: DeviceInfoMessage[];
  isGroupLeader?: boolean | undefined;
  isAirplayActive?: boolean | undefined;
  systemPodcastApplication?: string | undefined;
  senderDefaultGroupUID?: string | undefined;
  airplayReceivers: string[];
  linkAgent?: string | undefined;
  clusterID?: string | undefined;
  clusterLeaderID?: string | undefined;
  clusterType?: number | undefined;
  isClusterAware?: boolean | undefined;
  modelID?: string | undefined;
  supportsMultiplayer?: boolean | undefined;
  /** Example: 577E1BCA-2D9B-41C2-A8F8-C515CE8072D4 */
  routingContextID?: string | undefined;
  /** Example: 6E0ED14E-0BA4-4A84-AF30-9EB080E7AA73 */
  airPlayGroupID?: string | undefined;
  /** Example: com.apple.iBooks */
  systemBooksApplication?: string | undefined;
  clusteredDevices: DeviceInfoMessage[];
  parentGroupContainsDiscoverableGroupLeader?: number | undefined;
  groupContainsDiscoverableGroupLeader?: number | undefined;
  lastKnownClusterType?: number | undefined;
  /** repeated ... allClusteredDevices = 48; */
  supportsOutputContextSync?: boolean | undefined;
  computerName?: string | undefined;
  configuredClusterSize?: number | undefined;
  preferredEncoding?: PreferredEncoding_Enum | undefined;
  _unknownFields?: { [key: number]: Uint8Array[] } | undefined;
}