// src/protocols/mrp/generated/ClientUpdatesConfigMessage.ts
export interface ClientUpdatesConfigMessage {
  artworkUpdates?: boolean | undefined;
  nowPlayingUpdates?: boolean | undefined;
  volumeUpdates?: boolean | undefined;
  keyboardUpdates?: boolean | undefined;
  outputDeviceUpdates?: boolean | undefined;
  _unknownFields?: { [key: number]: Uint8Array[] } | undefined;
}